:recycle: group record
