+++ src/record.rs
@@ -1,2 +1,3 @@
+pub mod group;
 pub mod title;
 pub mod user;
