+++ src/response/group/groups.rs
@@ -1,4 +1,4 @@
-use crate::response::group::GroupRecord;
+use crate::record::group::GroupRecord;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Clone, Serialize)]
